refactor: rename agent registration files to enrollment
