+++ src/agent/CMakeLists.txt
@@ -38,7 +38,7 @@ endif()
 
 list(APPEND SOURCES
     src/agent.cpp
-    src/agent_registration.cpp
+    src/agent_enrollment.cpp
     src/agent_runner.cpp
     src/command_handler_utils.cpp
     src/message_queue_utils.cpp
